Move `**/dtos/user.*.dto.js` to `**/dtos/user/user.*.dto.js`
