+++ src/models/dtos/user/user.join.dto.js
@@ -1,4 +1,4 @@
-import { UserEntity } from '../entity/_.export.js';
+import { UserEntity } from '../../entity/_.export.js';
 
 /**
  *
